cloud_function/analyze-job/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
deprecation==2.1.0
distro==1.9.0
Flask==3.0.0
functions-framework==3.5.0
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-secret-manager==2.17.0
google-cloud-vision==3.5.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
openai==1.8.0
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pypdf==3.17.4
requests==2.31.0
rsa==4.9
sniffio==1.3.0
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
watchdog==3.0.0
Werkzeug==3.0.1
